added logout and jwt claims

+++ resources/item.py
@@ -3,7 +3,7 @@ from flask_smorest import Blueprint, abort
 from schemas import ItemSchema, ItemUpdateSchema
 from sqlalchemy.exc import SQLAlchemyError
 from models import ItemModel
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from db import db 
 
 blp = Blueprint ("items", __name__, description="Operations on stores")
@@ -17,6 +17,9 @@ class Item(MethodView):
         return item
     @jwt_required() 
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin priviledge is required")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
